refactor(multiorch): extract leader-fitness selection into its own file
selectFittestLeader replaces the mutating orderByLeaderFitness sort with a
non-mutating selection over just the eligible-leader candidates and their
health (fitness never depends on the recruitment result's rule/revocation).
buildFailoverProposal now takes the already-chosen leader as a plain value
instead of deriving it from EligibleLeaders[0], decoupling proposal
construction from leader selection entirely.

postgresReadyLess, leadershipSignalPriority, and poolerHealthStateLess move
into the new leader_fitness.go alongside selectFittestLeader, consolidating
logic that was previously split across coordinator.go and rule_change.go.
