+++ go/services/multiorch/consensus/rule_change.go
@@ -373,22 +373,25 @@ func (r *coordinatorLedRuleChange) promote(
 	return err
 }
 
-// buildFailoverProposal constructs a CoordinatorProposal for normal failover.
-// It picks the first node from result.EligibleLeaders (already sorted by
-// Coordinator.runFailover: postgres readiness, then leadership signal) and
-// resolves its contact address from addressByID. Resigned poolers are
+// buildFailoverProposal constructs a CoordinatorProposal for normal failover,
+// given leader (already chosen by the caller — see selectFittestLeader) and
+// its contact address resolved from addressByID. Resigned poolers are
 // filtered out upstream in Coordinator.runFailover.
 func buildFailoverProposal(
 	result commonconsensus.RecruitmentResult,
+	leader *clustermetadatapb.ConsensusStatus,
 	addressByID map[string]*clustermetadatapb.PoolerAddress,
 ) (*consensusdatapb.CoordinatorProposal, error) {
 	if result.OutgoingRule == nil {
 		return nil, errors.New("no committed rule found; use bootstrap path for fresh clusters")
 	}
-	if len(result.EligibleLeaders) == 0 {
+	if leader == nil {
 		return nil, errors.New("no eligible leaders for failover proposal")
 	}
-	leader := result.EligibleLeaders[0]
+	// Not verified here that leader is actually a member of
+	// result.EligibleLeaders: validateProposal (called by both
+	// BuildSafeProposal and CheckProposalPossible right after this function
+	// returns) already rejects a proposal whose leader isn't eligible.
 
 	addr, ok := addressByID[topoclient.ClusterIDString(leader.GetId())]
 	if !ok {
@@ -413,54 +416,6 @@ func buildFailoverProposal(
 			},
 		},
 	}, nil
-}
-
-// leadershipSignalPriority maps each leadership signal to its sort priority.
-// Lower values sort first — nodes with higher priority values are deprioritised
-// as leader candidates when LSNs are tied. Add new signals here to extend the
-// ordering without touching poolerHealthStateLess.
-var leadershipSignalPriority = map[clustermetadatapb.LeadershipSignal]int{
-	clustermetadatapb.LeadershipSignal_LEADERSHIP_SIGNAL_UNKNOWN:             0,
-	clustermetadatapb.LeadershipSignal_LEADERSHIP_SIGNAL_ACTIVE:              0,
-	clustermetadatapb.LeadershipSignal_LEADERSHIP_SIGNAL_REQUESTING_DEMOTION: 1,
-}
-
-// poolerHealthStateLess returns a less function for sort.SliceStable that
-// orders ConsensusStatus entries by leadershipSignalPriority. It is used in
-// Coordinator.runFailover to prefer nodes with lower priority values among
-// candidates that share the highest LSN.
-//
-// WAL position is always the primary criterion: a node with a higher-priority
-// signal still wins if it holds a strictly higher LSN than every other node.
-// This tiebreaker only affects the ordering of tied eligible leaders.
-//
-// Recruited nodes participate in the outgoing-cohort quorum check regardless of
-// their leadership signal — this tiebreaker only affects which tied node is
-// proposed as leader, not the quorum denominator.
-func poolerHealthStateLess(healthByID map[string]*multiorchdatapb.PoolerHealthState) func(a, b *clustermetadatapb.ConsensusStatus) bool {
-	leadershipSignal := func(cs *clustermetadatapb.ConsensusStatus) clustermetadatapb.LeadershipSignal {
-		h := healthByID[topoclient.ClusterIDString(cs.GetId())]
-		return h.GetAvailabilityStatus().GetLeadershipStatus().GetSignal()
-	}
-	failoverSlotsReady := func(cs *clustermetadatapb.ConsensusStatus) int32 {
-		h := healthByID[topoclient.ClusterIDString(cs.GetId())]
-		return h.GetStatus().GetFailoverSlotsReady()
-	}
-	return func(a, b *clustermetadatapb.ConsensusStatus) bool {
-		sigA := leadershipSignal(a)
-		sigB := leadershipSignal(b)
-		if leadershipSignalPriority[sigA] != leadershipSignalPriority[sigB] {
-			return leadershipSignalPriority[sigA] < leadershipSignalPriority[sigB]
-		}
-		// Slot-aware tiebreak among otherwise-equal candidates: prefer the one
-		// with more failover-ready logical slots so a promotion keeps the most
-		// subscribers resumable (see the durable slot-creation barrier). This
-		// only reorders candidates that already tied on WAL position (the
-		// EligibleLeaders set) and on leadership signal, so it never trades data
-		// safety or a resign intent for slot readiness. Zero when slot-based
-		// replication is off, leaving the ordering unchanged.
-		return failoverSlotsReady(a) > failoverSlotsReady(b)
-	}
 }
 
 // buildBootstrapProposal constructs a CoordinatorProposal for initial leader
